Updated package name, command renamed to commands
