+++ src/test/data/Default/watchlist.txt
@@ -1,2 +1,4 @@
 First | []
+Second | [1, 2]
+Third | [3]
 Test | []
