+++ src/test/java/anichan/commands/AddToWatchlistCommandTest.java
@@ -1,4 +1,4 @@
-package anichan.command;
+package anichan.commands;
 
 import anichan.human.User;
 import org.junit.jupiter.api.BeforeEach;
