+++ src/test/java/anichan/commands/EstimateCommandTest.java
@@ -1,4 +1,4 @@
-package anichan.command;
+package anichan.commands;
 
 import anichan.anime.AnimeData;
 import anichan.exception.AniException;
