+++ src/test/java/anichan/commands/SearchCommandTest.java
@@ -1,4 +1,4 @@
-package anichan.command;
+package anichan.commands;
 
 import anichan.anime.AnimeData;
 import anichan.exception.AniException;
